chore(frontend): consolidate duplicated helpers (isMac, isWindows, shortcutLabel, formatFileSize, formatDuration) into utils/platform and utils/format
