+++ frontend/src/components/Layout/Sidebar.tsx
@@ -15,14 +15,10 @@ import {
 import { useUserGroupsWithChannels } from "../../hooks/queries/useGroups";
 import { useDMConversations } from "../../hooks/queries/useMessages";
 import { useAppStore } from "../../stores/appStore";
+import { shortcutLabel } from "../../utils/platform";
 
 const SIDEBAR_WIDTH = 220;
 const COLLAPSED_GROUPS_KEY = "pollis.sidebar.collapsedGroups";
-
-const isMac =
-  typeof navigator !== "undefined" &&
-  navigator.platform.toUpperCase().indexOf("MAC") >= 0;
-const TOGGLE_SHORTCUT_LABEL = isMac ? "⌘B" : "Ctrl+B";
 
 interface SidebarProps {
   isOpen: boolean;
@@ -226,8 +222,8 @@ export const Sidebar: React.FC<SidebarProps> = ({ isOpen, onToggle }) => {
         type="button"
         data-testid="sidebar-close"
         onClick={onToggle}
-        aria-label={`Close sidebar (${TOGGLE_SHORTCUT_LABEL})`}
-        title={`Close sidebar (${TOGGLE_SHORTCUT_LABEL})`}
+        aria-label={`Close sidebar (${shortcutLabel("B")})`}
+        title={`Close sidebar (${shortcutLabel("B")})`}
         style={{
           flexShrink: 0,
           display: "flex",
@@ -263,7 +259,7 @@ export const Sidebar: React.FC<SidebarProps> = ({ isOpen, onToggle }) => {
             lineHeight: 1.2,
           }}
         >
-          {TOGGLE_SHORTCUT_LABEL}
+          {shortcutLabel("B")}
         </kbd>
       </button>
     </aside>
