+++ frontend/src/components/Message/MessageItem.tsx
@@ -4,6 +4,7 @@ import { save as saveDialog } from "@tauri-apps/plugin-dialog";
 import { writeFile } from "@tauri-apps/plugin-fs";
 import { Reply, Download, Film, Check, Edit2, Trash2 } from "lucide-react";
 import { getFileIcon } from "../../utils/fileIcon";
+import { formatFileSize, formatDuration } from "../../utils/format";
 import { useAppStore } from "../../stores/appStore";
 import { downloadAndDecryptMedia, getMediaUrl } from "../../services/r2-upload";
 import { LinkifiedText } from "../ui/LinkifiedText";
@@ -298,19 +299,6 @@ const BlurhashCanvas: React.FC<{ hash: string; width: number; height: number }> 
       />
     </div>
   );
-};
-
-const formatFileSize = (bytes: number) => {
-  if (bytes === 0) { return ""; }
-  const sizes = ["B", "KB", "MB", "GB"];
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  return `${parseFloat((bytes / Math.pow(1024, i)).toFixed(1))}${sizes[i]}`;
-};
-
-const formatDuration = (seconds: number): string => {
-  const m = Math.floor(seconds / 60);
-  const s = Math.floor(seconds % 60);
-  return `${m}:${String(s).padStart(2, "0")}`;
 };
 
 // Shared styles for lightbox action buttons (download / esc).
